Autosuggest renamed to AutoSuggestCommon
